refactor: 에러처리 수정

+++ src/domains/matches/matches.controller.ts
@@ -6,17 +6,16 @@ import {
   Post,
   Body,
   Put,
-  Patch,
   NotFoundException,
   ConflictException,
-  Logger,
+  ForbiddenException,
 } from '@nestjs/common';
 import { MatchesService } from './matches.service';
 import { User } from '@prisma/client';
 import { CreateMatchDto, RateDto, UpdateMatchDto } from './matches.dto';
 import { Private } from 'src/decorators/private.decorator';
 import { DAccount } from 'src/decorators/account.decorator';
-import { ALREADY_RATED, INVALID_MATCH, INVALID_RATING, SELF_RATING } from './matches-error.messages';
+import { ALREADY_RATED, INVALID_APPLICATION, INVALID_MATCH, INVALID_RATING, SELF_RATING, UNAUTHORIZED } from './matches-error.messages';
 import { PrismaService } from 'src/database/prisma/prisma.service';
 
 @Controller('matches')
@@ -33,6 +32,13 @@ export class MatchesController {
 
   @Get(':matchId')
   async findMatch(@Param('matchId') matchId: string) {
+    const match = await this.prismaService.match.findUnique({
+      where: { id: matchId },
+    });
+
+    if (!match) {
+      throw new NotFoundException(INVALID_MATCH);
+    }
     return await this.matchesService.findMatch(matchId);
   }
 
@@ -49,6 +55,15 @@ export class MatchesController {
     @Param('matchId') matchId: string,
     @Body() dto: UpdateMatchDto,
   ) {
+    const match = await this.prismaService.match.findUnique({
+      where: { id: matchId },
+    });
+    if (!match) {
+      throw new NotFoundException(INVALID_MATCH);
+    }
+    if (match.hostId !== user.id) {
+      throw new ForbiddenException(UNAUTHORIZED);
+    }
     return await this.matchesService.editMatch(user.id, matchId, dto);
   }
 
@@ -58,6 +73,16 @@ export class MatchesController {
     @DAccount('user') user: User,
     @Param('matchId') matchId: string,
   ) {
+    const match = await this.prismaService.match.findUnique({
+      where: { id: matchId },
+    });
+
+    if (!match) {
+      throw new NotFoundException(INVALID_MATCH);
+    }
+    if (match.hostId !== user.id) {
+      throw new ForbiddenException(UNAUTHORIZED);
+    }
     return await this.matchesService.deleteMatch(user.id, matchId);
   }
 
@@ -67,6 +92,20 @@ export class MatchesController {
     @DAccount('user') user: User,
     @Param('matchId') matchId: string,
   ) {
+    const match = await this.prismaService.match.findUnique({
+      where: { id: matchId },
+      include: {
+        participants: true,
+      },
+    });
+
+    if (!match) {
+      throw new NotFoundException(INVALID_MATCH);
+    }
+
+    if (match.hostId === user.id) {
+      throw new ConflictException(INVALID_APPLICATION)
+    }
     return await this.matchesService.participate(matchId, user.id);
   }
